requirements.txt
boto3==1.6.21
botocore==1.9.21
docutils==0.14
futures==3.2.0
jmespath==0.9.3
Pillow==5.0.0
python-dateutil==2.6.1
s3transfer==0.1.13
six==1.11.0
